Direct use of internal loaders
