+++ runtime/onert/core/include/loader/ModelLoader.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019 Samsung Electronics Co., Ltd. All Rights Reserved
+ * Copyright (c) 2024 Samsung Electronics Co., Ltd. All Rights Reserved
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,23 +14,29 @@
  * limitations under the License.
  */
 
-#ifndef __ONERT_LOADER_TFLITE_LOADER_H__
-#define __ONERT_LOADER_TFLITE_LOADER_H__
+#ifndef __ONERT_LOADER_LOADERS_H__
+#define __ONERT_LOADER_LOADERS_H__
 
-#include "loader/ILoader.h"
+#include "ir/Model.h"
+
+#include <memory>
 
 namespace onert
 {
 namespace loader
 {
 
-class TFLiteLoader : public ILoader
-{
-public:
-  std::unique_ptr<ir::Model> loadFromFile(const std::string &file_path) override;
-};
+/**
+ * @brief     Create custom loader and load model from file
+ * @param[in] filename  File path to load model from
+ * @param[in] type      Type of custom loader to create
+ * @return    Loaded model.
+ *
+ * @note  Throw exception if failed to load model
+ */
+std::unique_ptr<ir::Model> loadModel(const std::string &filename, const std::string &type);
 
 } // namespace loader
 } // namespace onert
 
-#endif // __ONERT_LOADER_TFLITE_LOADER_H__
+#endif // __ONERT_LOADER_LOADERS_H__
